--- a/examples/shadow.nim
+++ b/examples/shadow.nim
@@ -1,4 +1,4 @@
-import pixie, chroma, vmath
+import chroma, pixie, vmath
 
 let
   trees = readImage("examples/data/trees.png")
@@ -7,7 +7,7 @@ let
 image.fill(rgba(255, 255, 255, 255))
 
 var p: Path
-p.polygon(100, 100, 70, sides=8)
+p.polygon(100, 100, 70, sides = 8)
 p.closePath()
 
 var polyImage = newImage(200, 200)