--- a/examples/tiger.nim
+++ b/examples/tiger.nim
@@ -1,4 +1,4 @@
-import pixie, chroma, vmath
+import chroma, pixie, vmath
 
 let image = newImage(200, 200)
 image.fill(rgba(255, 255, 255, 255))