--- a/experiments/benchmark_svg_cairo.nim
+++ b/experiments/benchmark_svg_cairo.nim
@@ -1,4 +1,4 @@
-import svg_cairo, benchy
+import benchy, svg_cairo
 
 let data = readFile("tests/images/svg/Ghostscript_Tiger.svg")