--- a/src/pixie/fileformats/svg.nim
+++ b/src/pixie/fileformats/svg.nim
@@ -1,7 +1,7 @@
 ## Load and Save SVG files.
 
-import chroma, pixie/images, pixie/common, pixie/paths, vmath, xmlparser,
-  xmltree, strutils, strutils
+import chroma, pixie/common, pixie/images, pixie/paths, strutils, vmath,
+    xmlparser, xmltree
 
 const svgSignature* = "<?xml"