--- a/tests/benchmark_png.nim
+++ b/tests/benchmark_png.nim
@@ -1,5 +1,5 @@
-import pixie/fileformats/png, stb_image/read as stbi, stb_image/write as stbr,
-  benchy, nimPNG
+import benchy, nimPNG, pixie/fileformats/png, stb_image/read as stbi,
+    stb_image/write as stbr
 
 let data = readFile("tests/images/png/lenna.png")