--- a/tests/benchmark_svg.nim
+++ b/tests/benchmark_svg.nim
@@ -1,4 +1,4 @@
-import pixie/fileformats/svg, benchy
+import benchy, pixie/fileformats/svg
 
 let data = readFile("tests/images/svg/Ghostscript_Tiger.svg")