--- a/tests/fuzz_paths.nim
+++ b/tests/fuzz_paths.nim
@@ -1,4 +1,4 @@
-import random, strformat, pixie/paths, pixie/common, strformat
+import pixie/common, pixie/paths, random, strformat
 
 randomize()