--- a/tests/fuzz_svg.nim
+++ b/tests/fuzz_svg.nim
@@ -1,4 +1,4 @@
-import random, strformat, pixie/fileformats/svg, pixie/common
+import pixie/common, pixie/fileformats/svg, random, strformat
 
 randomize()