--- a/tests/test_paths.nim
+++ b/tests/test_paths.nim
@@ -1,4 +1,4 @@
-import pixie, chroma, pixie/fileformats/png
+import chroma, pixie, pixie/fileformats/png
 
 block:
   let pathStr = """