… experiments/trapezoids/g.png
… experiments/trapezoids/heart.png
… experiments/trapezoids/l.png
… pixie.nimble
… tests/benchmark_paths.nim