`nim c -d:release --debugger:native -f -r .\tests\benchmark_svg.nim`
before:
svg decode ........................ 23.111 ms 24.079 ms ±0.677 x207
after:
svg decode ........................ 22.295 ms     23.101 ms    ±0.446   x216

`nim c -d:release --debugger:native -r .\experiments\benchmark_cairo.nim`
before:
`pixie .............................. 0.425 ms      0.438 ms    ±0.020  x1000`
after:
`pixie .............................. 0.251 ms      0.259 ms    ±0.007  x1000`